+++ backend/canisters/group/impl/src/queries/summary.rs
@@ -23,7 +23,7 @@ fn summary_impl(on_behalf_of: Option<Principal>, state: &RuntimeState) -> Respon
     };
 
     if let Some(member) = state.data.get_member(caller) {
-        let summary = state.summary(member);
+        let summary = state.summary(&member);
         Success(SuccessResult { summary })
     } else {
         CallerNotInGroup